Update delegate urls argument for dashboard oauth (#412)
There was a misconfiguration in the --openshift-delegate-urls parameter
for the `authentication` overlay of the dashboard. This caused
authentication/authorization to fail for some use cases, particularly
when attempting to authenticate to the dashboard programmatically using
a bearer token. This change updates the argument to a valid value to fix
this use case.