[feat] update Notebook annotations - don't pass the logout url annotation
With the move to the Gateway API, we decided to use single domain name
for now. This means same domain for all Notebook instances and as such,
SSO works for all of them including the Dashboard. We decided to remove
the Log out functionality from the Notebook instances. This decision
may be revisited in the future if we decide to move to the multiple
domain approach.

For this change, we stop passing down the relevant annotation with the
URL used for the logout operation.

We also updated the arguments provided for the Jupyterlab app in the
workbench removing the `--ServerApp.tornado_settings` parameter. This
change will remove the presence of the `File -> Log Out` and `File ->
Shut Down` options from the Jupyterlab menu.