Implement MaaS tiers CRUD operations on BFF
This adds endpoints for the following operations on MaaS tiers:
* GET /api/v1/tiers - Fetch the list of tiers
* GET /api/v1//tier/:name - Fetch the data of a specific tier
* POST /api/v1/tier - Create a tier
* PUT /api/v1/tier/:name - Update a tier
* DELETE api/V1/tier/:name - Delete a tier

The _POST data_ (create/update endpoints) is expected in JSON format. Similarly, responses are in JSON format.

Although the endpoints expose and can accept data about policies (token/request rate limits) and models that belong to the tier, such data is not handled in this commit and fake data is returned. Only the main tier definition, and groups enrollment is properly handled.